Preliminary support for section comments and table of contents
A section comment is a documentation comment (the section title)
followed by another documentation comment (the section description).
This syntax is a bit crude and will (hopefully) be improved in the
future. For this reason, this feature is also not yet documented
in the README.md file.

Each documented library should contain at least one section comment.
Based on section comments, a table of contents is created.